requirements-worker.txt
Flask==3.0.2
Pillow==10.2.0
requests==2.31.0 